+++ tests/test_enumerate.py
@@ -54,6 +54,13 @@ def test_match_tags(context):
     raise NotImplementedError()
 
 
+def test_match_is_initialized(context):
+    funcname = 'udev_enumerate_add_match_is_initialized'
+    with pytest.patch_libudev(funcname) as func:
+        context.list_devices().match_is_initialized()
+        assert func.called
+
+
 def test_combined_matches_of_same_type(context):
     """
     Test for behaviour as observed in #1